ci(test): prevent wrong branch job
The job should protected us from mergeing into the wrong branch. The expected target branch for any feature branch, is the MAJOR_COMMIT_REF_NAME. It is defined in the ci-config.yaml. And inhereted from the major commit branch, when branching. This job expects branches to be merged into there upstream branch.